Refactor project structure: move types to models directory and update imports; add API DTOs and goal model
