+++ src/pages/exercises.astro
@@ -2,7 +2,7 @@
 import Layout from "../layouts/Layout.astro";
 import { mockExerciseTemplates } from "@/lib/mocks";
 import { ExerciseCatalog } from "@/components/exercises";
-import type { Category } from "@/types";
+import type { Category } from "@/models";
 
 const categories: Category[] = [
   { id: "strength", label: "Siłowe", icon: "💪" },
